Added Identity Server

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -5,6 +5,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -47,13 +48,13 @@ public class AuctionsController : ControllerBase
         return _mapper.Map<AuctionDto>(auction);
     }
 
+    [Authorize]
     [HttpPost]
     public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDTO auctionDTO)
     {
         var auction = _mapper.Map<Auction>(auctionDTO);
 
-        // TODO: add current user as seller
-        auction.Seller = "Test";
+        auction.Seller = User.Identity.Name;
 
         _context.Auctions.Add(auction);
 
@@ -69,6 +70,7 @@ public class AuctionsController : ControllerBase
         return CreatedAtAction(nameof(GetAuctionById), new {auction.Id}, newAuction);
     }
 
+    [Authorize]
     [HttpPut("{id}")]
     public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDTO updateAuctionDTO)
     {
@@ -77,7 +79,8 @@ public class AuctionsController : ControllerBase
         
         if(auction is null) return NotFound();
 
-        // TODO: Check seller = username
+        if(auction.Seller != User.Identity.Name) 
+            return Forbid();
 
         auction.Item.Make = updateAuctionDTO.Make ?? auction.Item.Make;
         auction.Item.Model = updateAuctionDTO.Model ?? auction.Item.Model;
@@ -93,13 +96,15 @@ public class AuctionsController : ControllerBase
         return BadRequest("Problem saving changes");
     }
 
+    [Authorize]
     [HttpDelete("{id}")]
     public async Task<ActionResult> DeleteAuction(Guid id)
     {
         var auction = await _context.Auctions.FindAsync(id);
         if(auction is null) return NotFound();
 
-        // TODO: Check seller = username
+        if(auction.Seller != User.Identity.Name)
+            return Forbid();
 
         _context.Auctions.Remove(auction);
